requirements.txt
django
tensorflow-gpu==1.2
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
django-form-utils